+++ app/pb_kits/playbook/packs/examples.js
@@ -27,6 +27,7 @@ import * as Currency from 'pb_currency/docs'
 import * as DashboardValue from 'pb_dashboard_value/docs'
 import * as Date from 'pb_date/docs'
 import * as DateRangeInline from 'pb_date_range_inline/docs'
+import * as DateRangeStacked from 'pb_date_range_stacked/docs'
 import * as DateStacked from 'pb_date_stacked/docs'
 import * as DateYearStacked from 'pb_date_year_stacked/docs'
 import * as DistributionBar from 'pb_distribution_bar/docs'
@@ -95,6 +96,7 @@ WebpackerReact.setup({
   ...DashboardValue,
   ...Date,
   ...DateRangeInline,
+  ...DateRangeStacked,
   ...DateStacked,
   ...DateYearStacked,
   ...DistributionBar,